add some helper functions for copying between virtual and physical memory
this was already possible, but now it's slightly safer

+++ src/kernel/mem/virt.c
@@ -41,7 +41,8 @@ bool virt_iter_next(struct virt_iter *iter) {
 		}
 	} else {
 		// "iterate" over physical memory
-		iter->frag = (void*) iter->_virt;
+		// the double cast supresses the warning about changing address spaces
+		iter->frag = (void*)(uintptr_t)iter->_virt;
 	}
 
 	iter->frag_len    = partial;
